LIBILS-176. Moved example code into "examples" directory.
https://issues.umd.edu/browse/LIBILS-176
